feat: 메인페이지 데이터 뿌리기 및 .jsx -> .tsx 리팩토링
