+++ package.json
@@ -14,7 +14,9 @@
     "@types/react": "^18.0.27",
     "@types/react-dom": "^18.0.10",
     "@types/react-router-dom": "^5.3.3",
+    "@types/react-youtube": "^7.10.0",
     "@types/styled-components": "^5.1.26",
+    "@types/swiper": "^6.0.0",
     "axios": "^1.2.3",
     "firebase": "^9.16.0",
     "react": "^18.2.0",
